requirements.txt
torch>=2.0.0
transformers>=4.30.0
easyocr>=1.7.0
pdf2image>=1.16.3
Pillow>=9.0.0
numpy>=1.21.0 